implemented O2 standard format, and table maker can use it

+++ BCLabManagerV2/TableMaker/Model/AndroidDriverMaker.cs
@@ -47,7 +47,7 @@ namespace BCLabManager
 
         //Initialize content of H file, currently using hard coding in code, but hopely we can read it from file, a sample file in particular folder
 
-        private static List<TableMakerProduct> GenerateCHFiles(Stage stage, ref UInt32 uErr, string OutFolder, string strCFileStandardName, string strHFileStandardName, List<string> strHHeaderComments, List<string> strCHeaderComments, List<int> ilstOCVVolt, List<int> voltList, List<float> listfTemp, List<float> listfCurr, List<List<int>> outYValue, double fCTABase, double fCTASlope)
+        private static List<TableMakerProduct> GenerateCHFiles(Stage stage, ref UInt32 uErr, string OutFolder, string strCFileStandardName, string strHFileStandardName, List<string> strHHeaderComments, List<string> strCHeaderComments, List<int> ilstOCVVolt, List<int> voltList, List<double> listfTemp, List<double> listfCurr, List<List<int>> outYValue, double fCTABase, double fCTASlope)
         {
             string localcFilePath = System.IO.Path.Combine(OutFolder, strCFileStandardName);
             string localhFilePath = System.IO.Path.Combine(OutFolder, strHFileStandardName);
@@ -85,7 +85,7 @@ namespace BCLabManager
 
             return output;
         }
-        private static List<string> GetCFileContent(string strCFileStandardName, string strHFileStandardName, List<string> strHHeaderComments, List<int> ilstOCVVolt, List<int> voltList, List<float> listfTemp, List<float> listfCurr, List<List<int>> outYValue, double fCTABase, double fCTASlope)
+        private static List<string> GetCFileContent(string strCFileStandardName, string strHFileStandardName, List<string> strHHeaderComments, List<int> ilstOCVVolt, List<int> voltList, List<double> listfTemp, List<double> listfCurr, List<List<int>> outYValue, double fCTABase, double fCTASlope)
         {
             string line = "";
             List<string> output = new List<string>();
@@ -231,7 +231,7 @@ namespace BCLabManager
         }
 
         //Initialize content of H file, currently using hard coding in code, but hopely we can read it from file, a sample file in particular folder
-        private static List<string> GetHFileContent(string strStandardH, List<string> strHHeaderComments, List<int> ilstOCVVolt, List<int> voltList, List<float> listfCurr, List<float> listfTemp, double fCTABase, double fCTASlope)
+        private static List<string> GetHFileContent(string strStandardH, List<string> strHHeaderComments, List<int> ilstOCVVolt, List<int> voltList, List<double> listfCurr, List<double> listfTemp, double fCTABase, double fCTASlope)
         {
             int iLineCmtHCFile = 4;
             int i = 0;
