+++ BCLabManagerV2/TableMaker/Model/AndroidModel.cs
@@ -8,8 +8,8 @@ namespace BCLabManager.Model
         public List<int> iOCVVolt { get; internal set; }
         public double fCTABase { get; internal set; }
         public double fCTASlope { get; internal set; }
-        public List<float> listfCurr { get; internal set; }
-        public List<float> listfTemp { get; internal set; }
+        public List<double> listfCurr { get; internal set; }
+        public List<double> listfTemp { get; internal set; }
         public List<List<int>> outYValue { get; internal set; }
     }
 }
