+++ BCLabManagerV2/TableMaker/Model/LiteModel.cs
@@ -8,7 +8,7 @@ namespace BCLabManager.Model
         public List<short> ilistCurr { get; set; }
         public List<double[]> flstdbDCapCof { get; set; }
         public List<double[]> flstdbKeodCof { get; set; }
-        public List<float> flstTblOCVCof { get; set; }
+        public List<double> flstTblOCVCof { get; set; }
         public List<string> FileNames { get; set; }
         public UInt32 uEoDVoltage { get; set; }
     }
