+++ BCLabManagerV2/TableMaker/Model/StandardDriverMaker.cs
@@ -43,7 +43,7 @@ namespace BCLabManager
             TableMakerService.InitializeHeaderInfor(ref uErr, project.BatteryType.Manufacturer, project.BatteryType.Name, project.AbsoluteMaxCapacity.ToString(), project.LimitedChargeVoltage.ToString(), project.CutoffDischargeVoltage.ToString(), type_id.ToString(), out strCHeaderComments);
             return GenerateCHFiles(stage, ref uErr, OutFolder, strFilePaths[0], strFilePaths[1], strHHeaderComments, strCHeaderComments, standardModel.iOCVVolt, VoltagePoints, standardModel.listfTemp, standardModel.listfCurr, standardModel.outYValue, standardModel.fCTABase, standardModel.fCTASlope);
         }
-        private static List<TableMakerProduct> GenerateCHFiles(Stage stage, ref UInt32 uErr, string OutFolder, string strCFileStandardName, string strHFileStandardName, List<string> strHHeaderComments, List<string> strCHeaderComments, List<int> ilstOCVVolt, List<int> voltList, List<float> listfTemp, List<float> listfCurr, List<List<int>> outYValue, double fCTABase, double fCTASlope)
+        private static List<TableMakerProduct> GenerateCHFiles(Stage stage, ref UInt32 uErr, string OutFolder, string strCFileStandardName, string strHFileStandardName, List<string> strHHeaderComments, List<string> strCHeaderComments, List<int> ilstOCVVolt, List<int> voltList, List<double> listfTemp, List<double> listfCurr, List<List<int>> outYValue, double fCTABase, double fCTASlope)
         {
             string localCFilePath = System.IO.Path.Combine(OutFolder, strCFileStandardName);
             string localHFilePath = System.IO.Path.Combine(OutFolder, strHFileStandardName);
@@ -83,7 +83,7 @@ namespace BCLabManager
         }
 
 
-        private static List<string> GetCFileContent(string strCFileStandardName, string strHFileStandardName, List<string> strHHeaderComments, List<int> ilstOCVVolt, List<int> voltList, List<float> listfTemp, List<float> listfCurr, List<List<int>> outYValue, double fCTABase, double fCTASlope)
+        private static List<string> GetCFileContent(string strCFileStandardName, string strHFileStandardName, List<string> strHHeaderComments, List<int> ilstOCVVolt, List<int> voltList, List<double> listfTemp, List<double> listfCurr, List<List<int>> outYValue, double fCTABase, double fCTASlope)
         {
             string line = "";
             List<string> output = new List<string>();
@@ -247,7 +247,7 @@ namespace BCLabManager
         }
 
 
-        private static List<string> GetHFileContent(string strStandardH, List<string> strHHeaderComments, List<int> ilstOCVVolt, List<int> voltList, List<float> listfCurr, List<float> listfTemp, double fCTABase, double fCTASlope)
+        private static List<string> GetHFileContent(string strStandardH, List<string> strHHeaderComments, List<int> ilstOCVVolt, List<int> voltList, List<double> listfCurr, List<double> listfTemp, double fCTABase, double fCTASlope)
         {
             int iLineCmtHCFile = 4;
             int i = 0;
